Decouple PhotoGallery component
