+++ src/app/galeria/page.tsx
@@ -1,34 +1,6 @@
 import Container from "@/components/shared/Container";
 import { Breadcrumbs } from "@/components/shared/Breadcrumbs/Breadcrumbs";
-import { Heading } from "@/components/shared/Heading/Heading";
-
-interface Props {
-  title: string;
-  date: string;
-}
-
-const PhotoGallery = ({ title, date }: Props) => (
-  <article className="w-full">
-    <header className="space-y-4">
-      <Heading
-        color="green"
-        contrast="yellow"
-        className="text-2xl tablet:text-26 desktop:text-28"
-        variant="h3"
-      >
-        {title}
-      </Heading>
-      <p className="font-lora text-2xl text-greenMain">{date}</p>
-    </header>
-  </article>
-);
-
-const galleries = [
-  { title: "Zjazd lorem ipsum", date: "18 sierpnia, 2020" },
-  { title: "Zjazd lorem ipsum", date: "18 sierpnia, 2020" },
-  { title: "Zjazd lorem ipsum", date: "18 sierpnia, 2020" },
-  { title: "Zjazd lorem ipsum", date: "18 sierpnia, 2020" },
-];
+import { PhotoGallery } from "./_components/PhotoGallery";
 
 export default function GalleryPage() {
   return (
@@ -49,3 +21,10 @@ export default function GalleryPage() {
     </div>
   );
 }
+
+const galleries = [
+  { title: "Zjazd lorem ipsum", date: "18 sierpnia, 2020" },
+  { title: "Zjazd lorem ipsum", date: "18 sierpnia, 2020" },
+  { title: "Zjazd lorem ipsum", date: "18 sierpnia, 2020" },
+  { title: "Zjazd lorem ipsum", date: "18 sierpnia, 2020" },
+];
